agent-orchestrator/requirements.txt
fastapi==0.109.2
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
azure-cosmos==4.6.0
azure-identity==1.25.2
pydantic[email]==1.10.14
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.26.0
authlib==1.3.0
